Crate bp_messages

Primitives of messages module.

Modules

source_chain

Primitives of messages module, that are used on the source chain.

target_chain

Primitives of messages module, that are used on the target chain.

Structs

DeliveredMessages

Delivered messages with their dispatch result.

InboundLaneData

Inbound lane data.

Message

Message as it is stored in the storage.

MessageData

Message data as it is stored in the storage.

MessageDetails

Message details, returned by runtime APIs.

MessageKey

Message key (unique message identifier) as it is stored in the storage.

OutboundLaneData

Outbound lane data.

UnrewardedRelayer

Unrewarded relayer entry stored in the inbound lane data.

UnrewardedRelayersState

Gist of InboundLaneData::relayers field used by runtime APIs.

Enums

OperatingMode

Messages pallet operating mode.

Traits

Parameter

Messages pallet parameter.

Functions

total_unrewarded_messages

Returns total number of messages in the InboundLaneData::relayers vector.

Type Definitions

BridgeMessageId

Message id as a tuple.

DispatchResultsBitVec

Bit vector of message dispatch results.

LaneId

Lane identifier.

MessageNonce

Message nonce. Valid messages will never have 0 nonce.

MessagePayload

Opaque message payload. We only decode this payload when it is dispatched.

Weight

Numeric range of a transaction weight.