bp_messages::source_chain

Trait DeliveryConfirmationPayments

pub trait DeliveryConfirmationPayments<AccountId, LaneId> {
    type Error: Debug + Into<&'static str>;

    // Required method
    fn pay_reward(
        lane_id: LaneId,
        messages_relayers: VecDeque<UnrewardedRelayer<AccountId>>,
        confirmation_relayer: &AccountId,
        received_range: &RangeInclusive<MessageNonce>,
    ) -> MessageNonce;
}

Manages payments that are happening at the source chain during delivery confirmation transaction.

Required Associated Types

type Error: Debug + Into<&'static str>

Error type.

Required Methods

fn pay_reward( lane_id: LaneId, messages_relayers: VecDeque<UnrewardedRelayer<AccountId>>, confirmation_relayer: &AccountId, received_range: &RangeInclusive<MessageNonce>, ) -> MessageNonce

Pay rewards for delivering messages to the given relayers.

The implementation may also choose to pay reward to the confirmation_relayer, which is a relayer that has submitted delivery confirmation transaction.

Returns number of actually rewarded relayers.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.