Trait bp_messages::source_chain::LaneMessageVerifier

pub trait LaneMessageVerifier<Submitter, Payload, Fee> {
    type Error: Debug + Into<&'static str>;
    fn verify_message(
        submitter: &Sender<Submitter>, 
        delivery_and_dispatch_fee: &Fee, 
        lane: &LaneId, 
        outbound_data: &OutboundLaneData, 
        payload: &Payload
    ) -> Result<(), Self::Error>;
}

Lane message verifier.

Runtime developer may implement any additional validation logic over message-lane mechanism. E.g. if lanes should have some security (e.g. you can only accept Lane1 messages from Submitter1, Lane2 messages for those who has submitted first message to this lane, disable Lane3 until some block, …), then it may be built using this verifier.

Any fee requirements should also be enforced here.

Associated Types

type Error: Debug + Into<&'static str>

Error type.

Required methods

fn verify_message(
submitter: &Sender<Submitter>,
delivery_and_dispatch_fee: &Fee,
lane: &LaneId,
outbound_data: &OutboundLaneData,
payload: &Payload
) -> Result<(), Self::Error>

Verify message payload and return Ok(()) if message is valid and allowed to be sent over the lane.