Trait bp_messages::source_chain::MessageDeliveryAndDispatchPayment

pub trait MessageDeliveryAndDispatchPayment<AccountId, Balance> {
    type Error: Debug + Into<&'static str>;
    fn pay_delivery_and_dispatch_fee(
        submitter: &Sender<AccountId>, 
        fee: &Balance, 
        relayer_fund_account: &AccountId
    ) -> Result<(), Self::Error>;

    fn pay_relayers_rewards(
        lane_id: LaneId, 
        messages_relayers: VecDeque<UnrewardedRelayer<AccountId>>, 
        confirmation_relayer: &AccountId, 
        received_range: &RangeInclusive<MessageNonce>, 
        relayer_fund_account: &AccountId
    );
}

Message delivery payment. It is called as a part of submit-message transaction. Transaction submitter is paying (in source chain tokens/assets) for:

submit-message-transaction-fee itself. This fee is not included in the delivery_and_dispatch_fee and is withheld by the regular transaction payment mechanism;

message-delivery-transaction-fee. It is submitted to the target node by relayer;

message-dispatch fee. It is paid by relayer for processing message by target chain;

message-receiving-delivery-transaction-fee. It is submitted to the source node by relayer.

So to be sure that any non-altruist relayer would agree to deliver message, submitter should set delivery_and_dispatch_fee to at least (equivalent of): sum of fees from (2) to (4) above, plus some interest for the relayer.

Associated Types

type Error: Debug + Into<&'static str>

Error type.

Required methods

fn pay_delivery_and_dispatch_fee(
submitter: &Sender<AccountId>,
fee: &Balance,
relayer_fund_account: &AccountId
) -> Result<(), Self::Error>

Withhold/write-off delivery_and_dispatch_fee from submitter account to some relayers-fund account.

fn pay_relayers_rewards(
lane_id: LaneId,
messages_relayers: VecDeque<UnrewardedRelayer<AccountId>>,
confirmation_relayer: &AccountId,
received_range: &RangeInclusive<MessageNonce>,
relayer_fund_account: &AccountId
)

Pay rewards for delivering messages to the given relayers.

The implementation may also choose to pay reward to the confirmation_relayer, which is a relayer that has submitted delivery confirmation transaction.