Trait bp_messages::source_chain::TargetHeaderChain

pub trait TargetHeaderChain<Payload, AccountId> {
    type Error: Debug + Into<&'static str>;
    type MessagesDeliveryProof: Parameter + Size;
    fn verify_message(payload: &Payload) -> Result<(), Self::Error>;

    fn verify_messages_delivery_proof(
        proof: Self::MessagesDeliveryProof
    ) -> Result<(LaneId, InboundLaneData<AccountId>), Self::Error>;
}

Target chain API. Used by source chain to verify target chain proofs.

All implementations of this trait should only work with finalized data that can’t change. Wrong implementation may lead to invalid lane states (i.e. lane that’s stuck) and/or processing messages without paying fees.

Associated Types

type Error: Debug + Into<&'static str>

Error type.

type MessagesDeliveryProof: Parameter + Size

Proof that messages have been received by target chain.

Loading content...
Required methods

fn verify_message(payload: &Payload) -> Result<(), Self::Error>

Verify message payload before we accept it.

CAUTION: this is very important function. Incorrect implementation may lead to stuck lanes and/or relayers loses.

The proper implementation must ensure that the delivery-transaction with this payload would (at least) be accepted into target chain transaction pool AND eventually will be successfully ‘mined’. The most obvious incorrect implementation example would be implementation for BTC chain that accepts payloads larger than 1MB. BTC nodes aren’t accepting transactions that are larger than 1MB, so relayer will be unable to craft valid transaction => this (and all subsequent) messages will never be delivered.

fn verify_messages_delivery_proof(
proof: Self::MessagesDeliveryProof
) -> Result<(LaneId, InboundLaneData<AccountId>), Self::Error>

Verify messages delivery proof and return lane && nonce of the latest recevied message.

Loading content...