Module bp_messages::target_chain

Primitives of messages module, that are used on the target chain.

Structs

DispatchMessage

Message with decoded dispatch payload.

DispatchMessageData

Message data with decoded dispatch payload.

ForbidInboundMessages

Structure that may be used in place of MessageDispatch on chains, where inbound messages are forbidden.

FromBridgedChainMessagesProof

Messages proof from bridged chain.

ProvedLaneMessages

Proved messages from single lane of the source chain.

Traits

DeliveryPayments

Manages payments that are happening at the target chain during message delivery transaction.

MessageDispatch

Called when inbound message is received.

Type Aliases

ProvedMessages

Proved messages from the source chain.