Crate bp_messages

Primitives of messages module.

Modules

source_chain

Primitives of messages module, that are used on the source chain.

storage_keys

Storage keys of bridge messages pallet.

target_chain

Primitives of messages module, that are used on the target chain.

Structs

DeliveredMessages

Delivered messages with their dispatch result.

InboundLaneData

Inbound lane data.

InboundMessageDetails

Inbound message details, returned by runtime APIs.

LaneId

Lane id which implements TypeId.

Message

Message as it is stored in the storage.

MessageKey

Message key (unique message identifier) as it is stored in the storage.

OutboundLaneData

Outbound lane data.

OutboundMessageDetails

Outbound message details, returned by runtime APIs.

ReceivedMessages

Received messages with their dispatch result.

UnrewardedRelayer

Unrewarded relayer entry stored in the inbound lane data.

UnrewardedRelayersState

Gist of InboundLaneData::relayers field used by runtime APIs.

Weight

Enums

BridgeMessagesCall

A minimized version of pallet-bridge-messages::Call that can be used without a runtime.

MessagesOperatingMode

Messages pallet operating mode.

ReceptionResult

Result of single message receival.

VerificationError

Error that happens during message verification.

Constants

HARD_MESSAGE_SIZE_LIMIT

Hard limit on message size that can be sent over the bridge.

Traits

ChainWithMessages

Substrate-based chain with messaging support.

Functions

calc_relayers_rewards

Calculate the number of messages that the relayers have delivered.

maximal_incoming_message_size

Return maximal size of the message the chain with max_extrinsic_size is able to receive.

Type Aliases

BridgeMessageId

Message id as a tuple.

MessageNonce

Message nonce. Valid messages will never have 0 nonce.

MessagePayload

Opaque message payload. We only decode this payload when it is dispatched.