bp_messages
Module source_chain

Primitives of messages module, that are used on the source chain.

Structs

ForbidOutboundMessages

Structure that may be used in place of TargetHeaderChain, LaneMessageVerifier and MessageDeliveryAndDispatchPayment on chains, where outbound messages are forbidden.

NoopMessagesBridge

Bridge that does nothing when message is being sent.

SendMessageArtifacts

Send message artifacts.

Traits

DeliveryConfirmationPayments

Manages payments that are happening at the source chain during delivery confirmation transaction.

LaneMessageVerifier

Lane message verifier.

MessagesBridge

Messages bridge API to be used from other pallets.

OnMessagesDelivered

Callback that is called at the source chain (bridge hub) when we get delivery confirmation for new messages.

TargetHeaderChain

Target chain API. Used by source chain to verify target chain proofs.

Type Aliases

RelayersRewards

Number of messages, delivered by relayers.