Module bp_messages::source_chain

Primitives of messages module, that are used on the source chain.

Structs

ForbidOutboundMessages

Structure that may be used in place MessageDeliveryAndDispatchPayment on chains, where outbound messages are forbidden.

FromBridgedChainMessagesDeliveryProof

Messages delivery proof from the bridged chain.

SendMessageArtifacts

Send message artifacts.

Traits

DeliveryConfirmationPayments

Manages payments that are happening at the source chain during delivery confirmation transaction.

MessagesBridge

Messages bridge API to be used from other pallets.

OnMessagesDelivered

Callback that is called at the source chain (bridge hub) when we get delivery confirmation for new messages.

Type Aliases

RelayersRewards

Number of messages, delivered by relayers.