Trait bp_messages::source_chain::MessageDeliveryAndDispatchPayment

pub trait MessageDeliveryAndDispatchPayment<AccountId, Balance> {
    type Error: Debug + Into<&'static str>;
    fn pay_delivery_and_dispatch_fee(
        submitter: &Sender<AccountId>, 
        fee: &Balance, 
        relayer_fund_account: &AccountId
    ) -> Result<(), Self::Error>;

    fn pay_relayers_rewards(
        confirmation_relayer: &AccountId, 
        relayers_rewards: RelayersRewards<AccountId, Balance>, 
        relayer_fund_account: &AccountId
    );

    fn initialize(_relayer_fund_account: &AccountId) -> usize { ... }
}

Message delivery payment. It is called as a part of submit-message transaction. Transaction submitter is paying (in source chain tokens/assets) for:

submit-message-transaction-fee itself. This fee is not included in the delivery_and_dispatch_fee and is witheld by the regular transaction payment mechanism;

message-delivery-transaction-fee. It is submitted to the target node by relayer;

message-dispatch fee. It is paid by relayer for processing message by target chain;

message-receiving-delivery-transaction-fee. It is submitted to the source node by relayer.

So to be sure that any non-altruist relayer would agree to deliver message, submitter should set delivery_and_dispatch_fee to at least (equialent of): sum of fees from (2) to (4) above, plus some interest for the relayer.

Associated Types

type Error: Debug + Into<&'static str>

Error type.

Loading content...
Required methods

fn pay_delivery_and_dispatch_fee(
submitter: &Sender<AccountId>,
fee: &Balance,
relayer_fund_account: &AccountId
) -> Result<(), Self::Error>

Withhold/write-off delivery_and_dispatch_fee from submitter account to some relayers-fund account.

fn pay_relayers_rewards(
confirmation_relayer: &AccountId,
relayers_rewards: RelayersRewards<AccountId, Balance>,
relayer_fund_account: &AccountId
)

Pay rewards for delivering messages to the given relayers.

The implementation may also choose to pay reward to the confirmation_relayer, which is a relayer that has submitted delivery confirmation transaction.

Loading content...
Provided methods

fn initialize(_relayer_fund_account: &AccountId) -> usize

Perform some initialization in externalities-provided environment.

For instance you may ensure that particular required accounts or storage items are present. Returns the number of storage reads performed.

Loading content...