Module bp_messages::storage_keys

Storage keys of bridge messages pallet.

Constants

INBOUND_LANES_MAP_NAME

Name of the InboundLanes storage map.

OPERATING_MODE_VALUE_NAME

Name of the OPERATING_MODE_VALUE_NAME storage value.

OUTBOUND_LANES_MAP_NAME

Name of the OutboundLanes storage map.

OUTBOUND_MESSAGES_MAP_NAME

Name of the OutboundMessages storage map.

Functions

inbound_lane_data_key

Storage key of the inbound message lane state in the runtime storage.

message_key

Storage key of the outbound message in the runtime storage.

operating_mode_key

Storage key of the PalletOperatingMode value in the runtime storage.

outbound_lane_data_key

Storage key of the outbound message lane state in the runtime storage.