Trait bp_messages::target_chain::MessageDispatch

pub trait MessageDispatch<AccountId, Fee> {
    type DispatchPayload: Decode;
    fn dispatch_weight(
        message: &DispatchMessage<Self::DispatchPayload, Fee>
    ) -> Weight;

    fn dispatch(
        relayer_account: &AccountId, 
        message: DispatchMessage<Self::DispatchPayload, Fee>
    ) -> MessageDispatchResult;
}

Called when inbound message is received.

Associated Types

type DispatchPayload: Decode

Decoded message payload type. Valid message may contain invalid payload. In this case message is delivered, but dispatch fails. Therefore, two separate types of payload (opaque MessagePayload used in delivery and this DispatchPayload used in dispatch).

Required methods

fn dispatch_weight(
message: &DispatchMessage<Self::DispatchPayload, Fee>
) -> Weight

Estimate dispatch weight.

This function must: (1) be instant and (2) return correct upper bound of dispatch weight.

fn dispatch(
relayer_account: &AccountId,
message: DispatchMessage<Self::DispatchPayload, Fee>
) -> MessageDispatchResult

Called when inbound message is received.

It is up to the implementers of this trait to determine whether the message is invalid (i.e. improperly encoded, has too large weight, …) or not.

If your configuration allows paying dispatch fee at the target chain, then it must be paid inside this method to the relayer_account.